requirements.txt
django==4.2
Pillow==10.4.0
pytesseract==0.3.13